# perf: cache swc dependency analysis and don't hold onto `ParsedSource`s in memory