--- a/cli/deno_dir.rs
+++ b/cli/deno_dir.rs
@@ -58,6 +58,12 @@ impl DenoDir {
     self.root.join("lint_incremental_cache_v1")
   }
 
+  /// Path for caching swc dependency analysis.
+  pub fn dep_analysis_db_file_path(&self) -> PathBuf {
+    // bump this version name to invalidate the entire cache
+    self.root.join("dep_analysis_cache_v1")
+  }
+
   /// Path for the cache used for type checking.
   pub fn type_checking_cache_db_file_path(&self) -> PathBuf {
     // bump this version name to invalidate the entire cache