--- a/cli/module_loader.rs
+++ b/cli/module_loader.rs
@@ -65,10 +65,7 @@ impl CliModuleLoader {
     let found_url = graph_data.follow_redirect(specifier);
     match graph_data.get(&found_url) {
       Some(ModuleEntry::Module {
-        code,
-        media_type,
-        maybe_parsed_source,
-        ..
+        code, media_type, ..
       }) => {
         let code = match media_type {
           MediaType::JavaScript
@@ -90,11 +87,12 @@ impl CliModuleLoader {
           | MediaType::Jsx
           | MediaType::Tsx => {
             // get emit text
-            let parsed_source = maybe_parsed_source.as_ref().unwrap(); // should always be set
             emit_parsed_source(
               &self.ps.emit_cache,
+              &self.ps.parsed_source_cache,
               &found_url,
-              parsed_source,
+              *media_type,
+              code,
               &self.ps.emit_options,
               self.ps.emit_options_hash,
             )?
@@ -103,6 +101,10 @@ impl CliModuleLoader {
             panic!("Unexpected media type {} for {}", media_type, found_url)
           }
         };
+
+        // at this point, we no longer need the parsed source in memory, so free it
+        self.ps.parsed_source_cache.free(specifier);
+
         Ok(ModuleCodeSource {
           code,
           found_url,