This PR caches swc dependency analysis and changes it so that we don't hold onto `ParsedSource`s anymore while running a program. This means that memory usage is drastically reduced and we have a significant performance improvement because we no longer need to parse sources files on every run.

**Memory Usage**

Given the following code:

```ts
import { Application } from "https://deno.land/x/oak@v10.6.0/mod.ts";
import * as test from "https://deno.land/std@0.152.0/node/module_all.ts";
import chalk from "npm:chalk";

console.log(Application);
console.log(chalk);
console.log(test);

setInterval(() => console.log(5), 5000);
```

Before: 185MB
After: 39MB

**Speed**

In the example above, but removing the `setInterval`:

Before (after first run): ~940ms
After (after first run): ~460ms (removing dep analysis cache, brings it up to 970ms on first run)

For something like just importing ts-morph:

```ts
import { Project } from "https://deno.land/x/ts_morph/mod.ts";

console.log(Project);
```

Before (after first run): ~1080ms
After (after first run): ~225ms